Changes by Diego
Hi guys! 

First of all, I think that you have made a great job at making use of the descriptive statistics that we covered in class. I have only added a couple of titles for some graphs as well as the code chunk to get png files for each of the graphs you created. Moreover, when it comes to the measures of dispersion, I have also added a loop to calculate the variance. 

On the other hand, I think that you offer really good interpretations of the statistics that you use. Nevertheless, you have multiple markdown and readme documents. It would probably be easier to follow if you combine everything in one markdown document (this is just a suggestion, of course you can do however you think it is best). 

Best,
Diego